Move kitchen to ansible directory
kitchen-ansible doesn't really support the setup we have of running kitchen from
outside of the ansible directory. For example, none of our group/host variables
are being copied to the docker instance and have to be specified in the kitchen
file's `extra_vars` property. Or, if not specified, we might end up with
a default that doesn't match what we're actually using in production. Best to
use the existing group/host vars and run kitchen-ansible the way it wants.