Revert to the manual install of pycsw
Installing via `pip install -e` causes pip to pull in dependencies listed in
setup.py, which normally would be good, but in this case pulls in a more recent
version of SQLAlchemy, which breaks ckan.

Going back to the old behavior is just a work around. The right fix would be:
1. Move pycsw to its own virtualenv, separate from ckan.
1. Use pinned dependencies via requirements.txt or Pipfile.lock to install pycsw
   and dependencies.