[catalog] increase number of CKAN workers
In response to the outage on 2020-01-27, we think the catalog hosts have more
capacity to respond to requests. The theory is that Apache2 is queuing up
requests because the CKAN workers are all busy. There is plenty of CPU and
memory available on these hosts to accommodate additional CKAN processes.