Fix writing .retry files in container
app user doesn't have write permissions to the ansible directory, preventing it
from writing .retry files. This doesn't really matter, because we're not using
the .retry files and if we're trying to write the .retry file, the play has
already failed. This is more just to prevent the error message which is
distracting.