Fix CI
Adds ssh to the docker image. The porting guide to [2.8][1] wasn't clear on
this. It seems like the docker image in CI used to fallback to paramiko in the absence
of ssh. That doesn't seem to be the case with 2.8.

Since in staging/production we run with ssh, this also improves parity in those
environments.

[1]: https://docs.ansible.com/ansible/devel/porting_guides/porting_guide_2.8.html