[catalog-next] rename ansible groups
This is a bit of a hack. All group names (besides `all`) have the same variable
precedence. This means that if you define a variable `foo` in both catalog-admin and
catalog-web, the last one wins. We can work around this by using group names
that match the hierarchy when sorted alphabetically e.g. catalog-next ->
catalog-next-web -> catalog-next-web-admin.