ansible-inventory-diff tool
This is a work in progress but has been helpful for analyzing changes to the
Ansible inventory. This wraps git and ansible-inventory to generate a JSON
representation of the Ansible inventory at two different commits and then diff
them.

What you get is a diff of any changes to group assignments and variables, making
it much easier to spot when a variable is being inherited incorrectly.