[catalog-next] add max-requests
Recycle gunicorn workers every 5000-6000 requests. These numbers are a rough
guess to start. Previously this is unset, so workers are never recycled.

Assuming we want to recycle processes every ~60 minutes, with a total,
consistent site load of 2000 req/min and 4 web instances each with 6 processes
(on production):

    60m * 2000r/m / (4*6) = 5000

And then we use a 20% jitter to stagger restarts, effectively restarting every
60-72 minutes assuming the load never changes.