[ci] update Jenkinsfile for multi-workflow
A bit awkward, but to maintain separation between sandbox and BSP environments,
we're running two separate instances of Jenkins but sharing the same pipeline
file.

The sandbox Jenkins will be considered the "sandbox" workflow and the BSP
Jenkins will be considered the "production" workflow. The workflows are similar
except that the production workflow deploys into the three BSP environments
sequentially, staging, mgmt, and production.