[jenkins] support branch builds
We've been hard-coding the branch to build per job through the Ansible
configuration. This allows the branch to be passed as a build parameter so that
staging can build both `release/*` and `master` branches.

This also makes things less confusing between CircleCI and Jenkins, where you
might be triggering a build from one branch on CircleCI but Jenkins uses
a different branch from the job configuration. This removes that quirk.