[ci] tag datagov-deploy container based on branch
Application builds should always use the latest datagov-deploy container from
the main/master branch.

Without tagging the containers by branch, it's possible that the latest
container to build was on a develop or feature branch. This would lead to using
development datagov-deploy code to deploy production applications. We want to
avoid that.