Use master as default branch
This updates documentation to drop `develop` as the default branch. All features
and bugfixes will be based on `master`. Hotfixes are no longer necessary, and it
will no longer be necessary to deploy to sandbox first.

`develop` branch becomes an ad-hoc un-protected branch that folks can use for
deploying to the sandbox for any development and testing.