[common] work-around issue with ipv6
The hardening playbook disables ipv6, but then attempts to configure it which
results in an error when sysctl reloads for an unrelated change. Instead, we
leave the ipv6 subsystem enabled, but disable it via sysctl.

This also cleans up the GRUB_CMDLINE_LINUX setting. The hardening playbook
creates multiple lines of this, which is not allowed. GRUB will only read the
last one which means some security settings have not been enabled correctly.
Work-around this by updating the last GRUB_CMDLINE_LINUX line with all the
options we expect.