[jenkins] remove disableRemotePolling
This was added as a work-around because the jobs did not seem to be detecting
changes (seeing an old SHA as being up-to-date). Unfortunately, using the
working directory for the polling causes some strange behavior. If we build
a branch (branch_name parameter), it seems like the polling uses that as the
current SHA which leads to changes being detected over and over.

In recent testing, polling seems to be working fine without
disableRemotePolling. Remove it for now and :fingers_crossed: